@@ -1,4 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:vlec2/details_page.dart';
+
+import 'data.dart';
 
 void main() => runApp(MyApp());
 
@@ -30,64 +33,110 @@ class _HomePageState extends State<HomePage> {
       appBar: AppBar(
         title: Text("Movie List"),
       ),
-      body: ListView.builder(itemBuilder: (BuildContext context, int i) {
-        if(i == 2) {
-          return Text("THIRD ELEMENT", style: TextStyle(fontSize: 40),);
-        }
-        String title = "Breaking Bad $i";
-        String genre = "Genre $i";
-        return MovieCard(title: title, genre: genre,);
-      }, itemCount: 1000,),
+      body: ListView.builder(
+        itemBuilder: (BuildContext context, int i) {
+          Movie movie = movieData[i];
+          return GestureDetector(
+            onTap: () {
+              // TODO: Navigate to detailsPage
+              Navigator.of(context)
+                  .push(MaterialPageRoute(builder: (context) => DetailsPage(movie: movieData[i],)));
+            },
+            child: MovieCard(
+              imgUrl: movie.imgUrl,
+              title: movie.title,
+              genre: movie.genre,
+              brief: movie.brief,
+            ),
+          );
+        },
+        itemCount: movieData.length,
+      ),
     );
   }
 }
 
 class MovieCard extends StatelessWidget {
-  final String title, genre;
+  final String title, genre, brief, imgUrl;
 
-  const MovieCard({Key key, @required this.title, @required this.genre}) : super(key: key);
+  const MovieCard(
+      {Key key,
+      @required this.imgUrl,
+      @required this.title,
+      @required this.genre,
+      @required this.brief})
+      : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     Size screenSize = MediaQuery.of(context).size;
     return Padding(
       padding: const EdgeInsets.only(top: 8.0, left: 8.0, right: 8.0),
-      child: Material(
-        elevation: 5,
-        borderRadius: BorderRadius.circular(10),
-        clipBehavior: Clip.antiAlias,
-        child: Row(
-          children: <Widget>[
-            Image.network(
-              HomePage.imgUrl,
-              width: screenSize.width / 2.7,
-              height: screenSize.height / 4,
-              fit: BoxFit.cover
-            ),
-            Expanded(
-              child: Padding(
-                padding: const EdgeInsets.only(left: 8.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: <Widget>[
-                    Text("$title", style: TextStyle(fontSize: 20, fontWeight: FontWeight.w600),),
-                    Text("$genre".toUpperCase(), style: TextStyle(fontSize: 19, fontWeight: FontWeight.w300),),
-                    SizedBox(height: 10,),
-                    Text("dolor sit amet Loremdolor sit amet Loremlor sit amet Lorem Ipsum dolor sit amet Lorem Ipsum dolor sit amet Lorem Ipsum dolor sit amet "),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.end,
-                      children: <Widget>[
-                        Icon(Icons.star, color: Colors.amber,),
-                        Icon(Icons.star, color: Colors.amber,),
-                        Icon(Icons.star, color: Colors.amber,),
-                        Icon(Icons.star_half, color: Colors.amber,),
-                        Icon(Icons.star_border, color: Colors.amber,)
-                      ],
-                    )
-                  ],
-                ),
+      child: Container(
+        height: screenSize.height / 4,
+        child: Material(
+          elevation: 5,
+          borderRadius: BorderRadius.circular(10),
+          clipBehavior: Clip.antiAlias,
+          child: Row(
+            children: <Widget>[
+              Hero(
+                tag: imgUrl,
+                child: Image.network(imgUrl,
+                    width: screenSize.width / 2.7,
+                    height: screenSize.height / 4,
+                    fit: BoxFit.cover),
               ),
-            )
-          ],
+              Expanded(
+                child: Padding(
+                  padding: const EdgeInsets.only(left: 8.0),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: <Widget>[
+                      Text(
+                        "$title",
+                        style: TextStyle(
+                            fontSize: 20, fontWeight: FontWeight.w600),
+                      ),
+                      Text(
+                        "$genre".toUpperCase(),
+                        style: TextStyle(
+                            fontSize: 19, fontWeight: FontWeight.w300),
+                      ),
+                      Text(
+                          "dolor sit amet Loremdolor sit amet Loremlor sit amet Lorem Ipsum dolor sit amet Lorem Ipsum dolor sit amet Lorem Ipsum dolor sit amet "),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.end,
+                        children: <Widget>[
+                          Icon(
+                            Icons.star,
+                            color: Colors.amber,
+                          ),
+                          Icon(
+                            Icons.star,
+                            color: Colors.amber,
+                          ),
+                          Icon(
+                            Icons.star,
+                            color: Colors.amber,
+                          ),
+                          Icon(
+                            Icons.star_half,
+                            color: Colors.amber,
+                          ),
+                          Icon(
+                            Icons.star_border,
+                            color: Colors.amber,
+                          )
+                        ],
+                      )
+                    ],
+                  ),
+                ),
+              )
+            ],
+          ),
         ),
       ),
     );
